(WIP) LSP: Support textDocument/prepareRename
'textDocument/prepareRename' can be used by the client to ask the
server the range of the symbol under the cursor which would be changed
by a subsequent call to 'textDocument/rename' with that position.

We can use this information to fill the prompt with an accurate prefill
which can improve the UX for renaming symbols when the symbol doesn't
align with the "word" textobject. (We currently use the "word"
textobject as a default value for the prompt.)

NOTE: This change is not complete. This change makes it impossible to
      rename symbols if prepareRename is not implemented or does not
      succeed. Before being used, this change should fall back to a
      default behavior of using the "word" textobject as the prefill.